Here's another test to answer the question: does JSONDecoder() do less work if you don't ask it to decode everything in the JSON?
The answer is a qualified yes. On my machine using Xcode 16:

Codable: 0.038sec / 27849kb transient heap allocations
JSONSerialization mapping to struct: 0.049s / 15408kb alloc
JSONSerialization without any mapping: 0.026s / 8785kb alloc
Partially decoded Codable: 0.007s / 5905kb alloc

The last item wraps airport10000 in an object with a couple little keys ("response": "ok", "status": "good", "payload": <3mb of airports>) and then decodes into a struct that only has the little keys.

It's quite a bit faster and uses less memory than just calling JSONSerialization and getting a dict back, which is great. You can see that JSONDecoder must not be calling JSONSerialization and operating on the results. If you need JSON parsing but can ignore a big part of your document, this might be a good choice.

On the other hand, the partial decode still uses ~6mb of memory -- copying its input then doing some work on it, perhaps. And it's taking 7ms to give results on the first 20 bytes or so of the document. So overall this is a terrible approach -- with a SAX parser, this should be doable in a few ns.